Create notes12-23-23
started taking notes on the 3 ways to write onclick events. this is not copy and pasted code.  may be mistakes

document.querySelector(".btn").addEventListener("click", function () } document.querySelector(".changedBox").style.display = "none"; })

or 

documentquerySelector(".btn").onclick = function () {

what i learned with the alerts first is that if you put the .onclick = alert("something");  the alert triggers before the onclick.  but if you put it in a function it's ok